Worked on dif_ekf_p1().
Turns out measurementFcn() needs to handle complex numbers; refactored it to do so.
Added zbound() helper to round really small values to zero. Matlab seems to do this automatically. We use zbound() to fix a problem in the measurementFcn() where there is a sqrt() calculation on a really small complex number. Normally, we can leave small numbers as is, but because of the sqrt(), the sign of the result is affected and that could have consequences on future multiplications.
Also discovered that the Matlab code works with seconds as the unit for timestamps.

+++ dkf/dkf_meas.h
@@ -31,25 +31,25 @@ namespace dkf
         double r_ij = 0;
         double R_ij = 0;
         // processed times
-        uint64_t T_rsp0 = 0;
-        uint64_t  T_rsp1 = 0;
-        uint64_t  T_rnd0 = 0;
-        uint64_t  T_rnd1 = 0;
+        double T_rsp0 = 0;
+        double  T_rsp1 = 0;
+        double  T_rnd0 = 0;
+        double  T_rnd1 = 0;
         // beaconing raw times
-        uint64_t t_bcn_tx = 0;
-        uint64_t t_bcn_rx = 0;
-        uint64_t t_i = 0;
-        uint64_t t_j = 0;
+        double t_bcn_tx = 0;
+        double t_bcn_rx = 0;
+        double t_i = 0;
+        double t_j = 0;
         // sequence number
         uint64_t seq = 0;
         // is this a queued message?
         int queued = 0;
-        uint64_t t0 = 0;
-        uint64_t t1 = 0;
-        uint64_t t2 = 0;
-        uint64_t t3 = 0;
-        uint64_t t4 = 0;
-        uint64_t t5 = 0;
+        double t0 = 0;
+        double t1 = 0;
+        double t2 = 0;
+        double t3 = 0;
+        double t4 = 0;
+        double t5 = 0;
 
         Meas() = default;
